Dashboard/PythonAPI/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
google-api-python-client